# internal/admin/openapi.yaml
openapi: "3.1.0"
info:
  title: tspages API
  description: Static site hosting platform for Tailscale networks.
  version: "1.0"
servers:
  - url: https://pages.{tailnet}
    description: Control plane
    variables:
      tailnet:
        default: example.ts.net
paths:
...
  /deploy/{site}/{id}/activate:
    post:
      operationId: activateDeployment
      summary: Activate a deployment
      description: Switches live traffic to a specific deployment. Useful for rollbacks.
      tags: [deploy]
      parameters:
        - $ref: "#/components/parameters/site"
        - $ref: "#/components/parameters/deploymentId"
      responses:
        "200":
          description: Deployment activated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentInfo"
        "404":
          description: Deployment not found or not complete.
      security:
        - tailscale: [deploy]
...
components:
  parameters:
    site:
      name: site
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$"
      description: Site name (DNS label).
    deploymentId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Deployment ID (8-character hex string).
...
  schemas:
...
    DeploymentInfo:
      type: object
      properties:
        id:
          type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        created_by_avatar:
          type: string
          format: uri
        size_bytes:
          type: integer
          format: int64
      required: [id, active]
...
  securitySchemes:
    tailscale:
      type: http
      scheme: mutual-tls
      description: |
        All requests are authenticated via Tailscale WhoIs. The caller's
        node identity is verified by the local Tailscale daemon. Access
        levels (view, deploy, admin) are granted via tailnet application
        grants with the configured capability name.